ml/web_app/requirements.txt
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
kaleido==0.2.1
MarkupSafe==1.1.1
numpy==1.20.2
pandas==1.2.4
plotly==4.14.3
python-dateutil==2.8.1
pytz==2021.1
retrying==1.3.3
scikit-learn==0.24.2
scipy==1.6.2
six==1.15.0
threadpoolctl==2.1.0
Werkzeug==1.0.1
mysql-connector-python==8.0.26